refactor: mvc 패턴으로 리팩토링
- 패키지명을 mvc 패턴에 따라 변경한다.
- Application에서 입출력(view)와 domain 메서드 호출 부분을 controller로 이전한다.
- InputValidator를 Validator로 이름을 수정하여 기능을 빠르게 파악할 수 있다.
- Validator를 controller 패키지로 이동하여 입력값 검증은 controller로 이전한다.
