+++ src/test/java/org/duckstudy/model/CarTest.java
@@ -1,4 +1,4 @@
-package org.duckstudy.movingcar;
+package org.duckstudy.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
